Rename file to match earlier naming convention
